requirements.txt
angrcli==1.2.0
astroid==3.3.5
asttokens==2.4.1
backports.strenum==1.3.1
bidict==0.23.1
binsync==4.0.0
bitarray==2.9.3
bitstring==4.2.3
cachetools==5.5.0
capstone==5.0.3
cart==1.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cmake==3.31.0.1
cmd2==2.5.6
comm==0.2.2
coverage==7.6.4
CppHeaderParser==2.7.4
debugpy==1.8.8
decorator==5.1.1
dill==0.3.9
distlib==0.3.9
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
flaky==3.8.1
future==1.0.0
ghidra-bridge==1.0.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.10
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.29.0
isort==5.13.2
itanium-demangler==1.1
jedi==0.19.1
jfx-bridge==1.0.0
JPype1==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keystone-engine==0.9.2
libbs==2.5.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mulpyplexer==0.9
nampa==0.1.1
nclib==1.0.5
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
nose==1.3.7
nose-timer==1.0.1
numpy==2.1.3
packaging==24.2
parso==0.8.4
patchelf-wrapper==1.2.0
pbr==6.1.0
pefile==2024.8.26
pexpect==4.9.0
platformdirs==4.3.6
plumbum==1.9.0
ply==3.11
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
pydot==3.0.2
pyelftools==0.31
pyformlang==1.0.10
pygdbmi==0.11.0.0
Pygments==2.18.0
pyhidra==1.3.0
pylint==3.3.1
pyparsing==3.2.0
pyperclip==1.9.0
pyqodeng==0.0.11
PySide6==6.8.0.2
PySide6-QtAds==4.3.1.1
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pyzmq==26.2.0
QtAwesome==1.3.1
QtPy==2.4.2
RapidFuzz==3.10.1
requests==2.32.3
rich==13.9.4
rpyc==6.0.1
shellphish-qemu==0.12.3
shiboken6==6.8.0.2
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
stack-data==0.6.3
stevedore==5.4.0
sympy==1.13.3
thefuzz==0.22.1
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
unicorn==2.0.1.post1
unique_log_filter==0.1.0
urllib3==2.2.3
virtualenv==20.28.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
wcwidth==0.2.13
z3-solver==4.13.0.0
